Update lucuma-core - moves EmailStatus to core
