+++ build.sbt
@@ -17,7 +17,7 @@ val logbackVersion             = "1.5.6"
 val log4catsVersion            = "2.6.0"
 val lucumaCatalogVersion       = "0.45.3"
 val lucumaItcVersion           = "0.21.8"
-val lucumaCoreVersion          = "0.96.2"
+val lucumaCoreVersion          = "0.96.3"
 val lucumaGraphQLRoutesVersion = "0.8.8"
 val lucumaSsoVersion           = "0.6.16"
 val munitVersion               = "0.7.29"
