+++ modules/service/src/main/scala/lucuma/odb/graphql/mapping/LeafMappings.scala
@@ -18,6 +18,7 @@ import io.circe.refined.*
 import io.circe.syntax.*
 import lucuma.core.data.EmailAddress
 import lucuma.core.enums.*
+import lucuma.core.enums.EmailStatus
 import lucuma.core.math.Epoch
 import lucuma.core.math.SignalToNoise
 import lucuma.core.model.CallForProposals
@@ -43,7 +44,6 @@ import lucuma.core.model.sequence.TimeChargeCorrection
 import lucuma.core.util.Timestamp
 import lucuma.odb.data.CallForProposalsType
 import lucuma.odb.data.EditType
-import lucuma.odb.data.EmailStatus
 import lucuma.odb.data.ExecutionEventType
 import lucuma.odb.data.Existence
 import lucuma.odb.data.Extinction
